pyproject.toml
[project]
name = "nonebot-plugin-bawiki"
version = "0.9.3"
description = "A nonebot2 plugin for Blue Archive."
authors = [{ name = "student_2333", email = "lgc2333@126.com" }]
dependencies = [
    "nonebot2>=2.0.0",
    "nonebot-adapter-onebot>=2.1.1",
    "nonebot-plugin-htmlrender>=0.2.0.1",
    "nonebot-plugin-apscheduler>=0.2.0",
    "beautifulsoup4>=4.11.1",
    "lxml>=4.9.1",
    "pil-utils>=0.1.4",
    "httpx>=0.24.1",
    "pypinyin>=0.49.0",
]
requires-python = ">=3.8,<4.0"
readme = "README.md"
license = { text = "MIT" }
keywords = ["blue archive", "nonebot", "nonebot2", "bot", "qq"]
classifiers = [
    "Natural Language :: Chinese (Simplified)",
    "Topic :: Communications",
    "Topic :: Communications :: Chat",
    "Topic :: Communications :: Chat :: ICQ",
    "Topic :: Games/Entertainment",
    "Topic :: Games/Entertainment :: Board Games",
    "Topic :: Games/Entertainment :: Turn Based Strategy",
    "Topic :: Games/Entertainment :: Side-Scrolling/Arcade Games",
    "Topic :: Internet",
    "Topic :: Internet :: WWW/HTTP",
    "Topic :: Internet :: WWW/HTTP :: Dynamic Content",
    "Topic :: Internet :: WWW/HTTP :: Dynamic Content :: Wiki",
]

[project.urls]
homepage = "https://github.com/lgc2333/nonebot-plugin-bawiki/"

[project.optional-dependencies]
menu = ["nonebot-plugin-PicMenu>=0.2"]

[tool.pdm.build]
includes = []

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"
